services/notification-service/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
redis==5.2.1
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0
httpx==0.28.1
aiosmtplib==5.1.0
opentelemetry-api==1.28.2
